Change: move function event to manage_events.c
